Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/browsers/history/the-location-interface/assign_after_load-expected.txt


FAIL Assignment to location after document is completely loaded assert_equals: expected 4 but got 3
